* Code cleanup. * Better error handling. * Better locking (as much as possible with us being a module). * Added unit tester (unit_test.py)

+++ useless.h
@@ -3,6 +3,6 @@
 struct useless_restore_req
 {
 	unsigned int pid;
-	unsigned int inode;
+	unsigned long inode;
 };
